requirements.txt
discord.py>=2.3.2
gspread>=5.11.3
oauth2client>=4.1.3
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib